working 2d array

+++ repeated_strings/Exercise.java
@@ -2,7 +2,7 @@ package repeated_strings;
 
 public class Exercise{
 
-    static void repeatedStrings(String string, long number){
+    static void repeatedStrings(String string, int number){
         int stringLength = string.length();
         int countOfAs = 0; // can use this for the end
         char[] stringArray = string.toCharArray(); // can use this to loop over string
@@ -11,7 +11,7 @@ public class Exercise{
                 countOfAs++;
             }
         }
-        long wholeRepeats = number / stringLength;
+        int wholeRepeats = number / stringLength;
         int remainingAs = 0;
         String remainigString = string.substring(0,number % stringLength);
         char[] remainingStringArray = remainigString.toCharArray();
@@ -28,3 +28,25 @@ public class Exercise{
         
     }
 }
+
+// char[] chrIn = s.toCharArray(); 
+// int i = 0; 
+// long max = 0; 
+// long quo = n / chrIn.length; 
+// long rem = n % chrIn.length; 
+// while (i < chrIn.length) { 
+//     if (chrIn[i] == 'a') { 
+//         max++; 
+//     } 
+//     i++; 
+// } 
+// i = 0; 
+// max = max * quo; 
+// while (i < rem) {
+//      if (chrIn[i] == 'a') { 
+//          max++; 
+//         } 
+//         i++;
+
+//     }
+//     return max;
